Replace bank cog with economy cog.
